+++ src/pso.c
@@ -23,11 +23,11 @@
 #define max(a, b) (((a) > (b)) ? (a) : (b))
 
   //returns swarm_size * dim size array
-void pso_rand_init(double* const population,
+void pso_rand_init(float* const population,
                    size_t swarm_size,
                    size_t dim,
-                   const double min_position,
-                   const double max_position) {
+                   const float min_position,
+                   const float max_position) {
   for (size_t particle=0; particle < swarm_size; particle++){
     for (size_t d=0; d<dim; d++){
       size_t idx = (particle*dim) + d;
@@ -39,7 +39,7 @@ void pso_rand_init(double* const population,
   //returns swarm_size size array
 void pso_eval_fitness(obj_func_t obj_func,
                       size_t swarm_size, size_t dim,
-                      const double* const positions, double* fitness) {
+                      const float* const positions, float* fitness) {
   for (size_t pop=0; pop<swarm_size; pop++){
     fitness[pop] = obj_func(positions+(pop*dim),dim);
   }
@@ -48,12 +48,12 @@ void pso_eval_fitness(obj_func_t obj_func,
 
   //returns swarm_size x dim size array
   // velocity at every posize_t of every particle
-void pso_gen_init_velocity(double* const velocity,
-                            const double* const positions,
+void pso_gen_init_velocity(float* const velocity,
+                            const float* const positions,
                             size_t swarm_size, size_t dim,
-                            const double min_position,
-                            const double max_position) {
-  double u[swarm_size*dim];
+                            const float min_position,
+                            const float max_position) {
+  float u[swarm_size*dim];
   pso_rand_init(u,swarm_size,dim,min_position,max_position);
   for (size_t particle=0; particle < swarm_size; particle++){
     for (size_t d=0; d<dim; d++){
@@ -65,12 +65,12 @@ void pso_gen_init_velocity(double* const velocity,
 
   // sorts the population in decreasing order of fitness /
   // i.e pop[0] posize_ts to best location yet.
-size_t pso_best_fitness(double* fitness,size_t dim, size_t swarm_size) {
+size_t pso_best_fitness(float* fitness,size_t dim, size_t swarm_size) {
   // find minimum of fitness
   // set index of mimima as minima_idx;
   // global_best = x+(dim*minima_idx);
   size_t minima_idx = 0;
-  double best_yet = fitness[0]; // global best at timestep t+1
+  float best_yet = fitness[0]; // global best at timestep t+1
   for (size_t p=1;p<swarm_size;p++){
     if (fitness[p]<best_yet) {
       best_yet = fitness[p];
@@ -82,21 +82,21 @@ size_t pso_best_fitness(double* fitness,size_t dim, size_t swarm_size) {
 }
 
   // limit max and min velocity values to ensure no boundary overflow
-void pso_generate_vel_limit(const double min_position,
-                            const double max_position,
-                            double* min_vel,
-                            double* max_vel) {
+void pso_generate_vel_limit(const float min_position,
+                            const float max_position,
+                            float* min_vel,
+                            float* max_vel) {
     *min_vel = min_position/VEL_LIMIT_SCALE;
     *max_vel = max_position/VEL_LIMIT_SCALE;
 }
 
   // updates velocities, works on array returned by gen_velocity
   //REMOVE VNEW WORK WITH VELOCITY
-void pso_update_velocity(double* velocity, double* positions,
-                         double* local_best_position,
-                         double* best, size_t swarm_size, size_t dim,
-                         const double min_vel,
-                         const double max_vel) {
+void pso_update_velocity(float* velocity, float* positions,
+                         float* local_best_position,
+                         float* best, size_t swarm_size, size_t dim,
+                         const float min_vel,
+                         const float max_vel) {
   for (size_t particle=0; particle < swarm_size; particle++){
     for (size_t d=0; d<dim; d++){
       size_t idx = (particle*dim) + d;
@@ -108,10 +108,10 @@ void pso_update_velocity(double* velocity, double* positions,
   return ;
 }
 
-void pso_update_position(double* positions, double* velocity,
+void pso_update_position(float* positions, float* velocity,
                          size_t swarm_size, size_t dim,
-                         const double min_position,
-                         const double max_position) {
+                         const float min_position,
+                         const float max_position) {
   for (size_t particle=0; particle < swarm_size; particle++){
     for (size_t d=0; d<dim; d++){
       size_t idx = (particle*dim) + d;
@@ -128,31 +128,31 @@ void pso_update_position(double* positions, double* velocity,
   /**
   Particle Swarm Algorithm to find the global minima of objective function using a nature mimicking swarm
   **/
-double * pso_basic(obj_func_t obj_func,
+float * pso_basic(obj_func_t obj_func,
                    size_t swarm_size,
                    size_t dim,
                    size_t max_iter,
-                   const double min_position,
-                   const double max_position) {
+                   const float min_position,
+                   const float max_position) {
   srand(100);
 
-  double min_vel = 0;
-  double max_vel = 0;
+  float min_vel = 0;
+  float max_vel = 0;
   pso_generate_vel_limit(min_position,max_position,&min_vel,&max_vel);
 
   size_t sizeof_position = dim*swarm_size;
   // randomly initialise postions of swarm particles
-  double current_positions[sizeof_position];
+  float current_positions[sizeof_position];
   pso_rand_init(current_positions,swarm_size, dim,min_position,max_position);
-  double local_best_position[sizeof_position];
-  memcpy(local_best_position,current_positions, sizeof_position*sizeof(double));
+  float local_best_position[sizeof_position];
+  memcpy(local_best_position,current_positions, sizeof_position*sizeof(float));
 
   // calculate the fitness at every position of the swarm
   size_t sizeof_fitness = swarm_size;
-  double current_fitness[sizeof_fitness];
+  float current_fitness[sizeof_fitness];
   pso_eval_fitness(obj_func,swarm_size,dim,current_positions,current_fitness);
-  double local_best_fitness[sizeof_fitness];
-  memcpy(local_best_fitness,current_fitness, sizeof_fitness*sizeof(double) );
+  float local_best_fitness[sizeof_fitness];
+  memcpy(local_best_fitness,current_fitness, sizeof_fitness*sizeof(float) );
 
   #ifdef DEBUG
       print_population(swarm_size, dim, current_positions); // printing the initial status of the population
@@ -160,13 +160,13 @@ double * pso_basic(obj_func_t obj_func,
       printf("# BEST FITNESS: %f\n", lowest_value(swarm_size, local_best_fitness));
   #endif
 
-  double p_velocity[sizeof_position];
+  float p_velocity[sizeof_position];
   pso_gen_init_velocity(p_velocity,current_positions,swarm_size,dim,min_position,max_position);
 
   size_t global_best_idx = pso_best_fitness(local_best_fitness,dim,swarm_size);
 
-  double* global_best_position = (local_best_position + (dim*global_best_idx));
-  double global_best_fitness = local_best_fitness[global_best_idx];
+  float* global_best_position = (local_best_position + (dim*global_best_idx));
+  float global_best_fitness = local_best_fitness[global_best_idx];
 
   size_t iter = 0;
   while ( iter < max_iter) {
@@ -204,8 +204,8 @@ double * pso_basic(obj_func_t obj_func,
 
   }
 
-  double* const best_solution = (double *const) malloc(dim);
-  memcpy(best_solution, global_best_position , dim*sizeof(double));
+  float* const best_solution = (float *const) malloc(dim);
+  memcpy(best_solution, global_best_position , dim*sizeof(float));
 
   return best_solution;
 }
